refactor(utils/validator.ts): 도메인이 있는 로직을 해당 도메인을 다루는 위치로 이동
