+++ src/components/feature/CardInfoFormSection/components/CardCVCInputField/utils.ts
@@ -1,0 +1,5 @@
+import { validateRange } from "@/utils/validator";
+
+export const validateCVCRange = (CVC: number) => {
+  return validateRange(CVC, 0, 999);
+};
