+++ src/components/feature/CardInfoFormSection/components/CardPasswordField/utils.ts
@@ -1,0 +1,5 @@
+import { validateRange } from "@/utils/validator";
+
+export const validatePasswordRange = (password: number) => {
+  return validateRange(password, 0, 99);
+};
